Move the Items get call to the UserInfo page
- this because the FavList needs the Items store obj populated
to load correctly and you can go to FavList before openning the
Items page
