+++ src/pages/UserInfo.js
@@ -1,11 +1,29 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
+import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import Navbar from '../components/Navbar';
+import Loader from '../components/Loader';
+import * as MyActions from '../actions';
 
 const UserInfo = props => {
-  const { user } = props;
+  const {
+    actions, items, fetchCall, user,
+  } = props;
+  const url = 'https://findmyitem-api.herokuapp.com/items';
+
+  useEffect(() => {
+    if (!items[0]) actions.fetchAPIcall(url, 'get', {});
+  }, []);
+
+  useEffect(() => actions.fetchAPIreset(), [items]);
+
+  useEffect(() => {
+    if (fetchCall.apiData) {
+      actions.setItems(fetchCall.apiData);
+    }
+  });
 
   return (
     <>
@@ -27,8 +45,14 @@ const UserInfo = props => {
               : 'Waiting or Not Found'
           }
         </ul>
-        <Link to="/items">See the Items</Link>
-        <Link to="/fav_list">See my Favorites List</Link>
+        {fetchCall.loading || !items[0]
+          ? <Loader />
+          : (
+            <>
+              <Link to="/items">See the Items</Link>
+              <Link to="/fav_list">See my Favorites List</Link>
+            </>
+          )}
       </section>
     </>
   );
@@ -36,8 +60,17 @@ const UserInfo = props => {
 
 UserInfo.propTypes = {
   user: PropTypes.objectOf(PropTypes.any).isRequired,
+  actions: PropTypes.objectOf(PropTypes.any).isRequired,
+  items: PropTypes.arrayOf(PropTypes.any).isRequired,
+  fetchCall: PropTypes.objectOf(PropTypes.any).isRequired,
 };
 
-const mapStateToProps = user => user;
+const mapStateToProps = ({ user, items, fetchCall }) => ({ user, items, fetchCall });
 
-export default connect(mapStateToProps)(UserInfo);
+function mapActionsToProps(dispatch) {
+  return {
+    actions: bindActionCreators({ ...MyActions }, dispatch),
+  };
+}
+
+export default connect(mapStateToProps, mapActionsToProps)(UserInfo);
